コンパイルエラー発生時のソースコードを別ウィンドウで表示できるようになりました．

+++ CocoViewer/src/src/coco/model/CCCompileError.java
@@ -37,7 +37,7 @@ public class CCCompileError {
 		return errorID;
 	}
 
-	public long getCorrectTime() {
+	public int getCorrectTime() {
 		return correctTime;
 	}
 }
